requirements.txt
# --- API & Server ---
fastapi==0.118.3
uvicorn==0.38.0
python-multipart==0.0.20
requests==2.32.4
python-dotenv==1.2.1

# --- OCR & Computer Vision ---
opencv-python-headless
Pillow==12.0.0
pytesseract==0.3.13

# --- PaddleOCR---
paddleocr>=3.0.0
paddlepaddle  # Use 'paddlepaddle-gpu' for gpu
shapely
pyclipper
imgaug
lmdb

# --- Machine Learning & Data ---
tensorflow
scikit-learn==1.6.1
langchain==0.1.20

